Move minified files in to Scripts Remove dist Update clean and .gitignore for new locations

+++ tasks/clean.js
@@ -1,10 +1,14 @@
 const fs = require("fs");
 const path = require("path");
 
-const distFolder = path.join(__dirname, "..", "dist");
+const scriptsFolder = path.join(__dirname, "..", "Scripts");
 
-const files = fs.readdirSync(distFolder);
+const files = fs.readdirSync(scriptsFolder);
 for (const file of files) {
-    const filePath = path.join(distFolder, file);
-    fs.unlinkSync(filePath);
+    console.log("Considering " + file);
+    if (file.match(/\.min\.js/)) {
+        console.log("  Deleting " + file);
+        const filePath = path.join(scriptsFolder, file);
+        fs.unlinkSync(filePath);
+    }
 }
